requirements.txt
aiofiles==0.8.0
aiogram==3.0.0b4
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
bidict==0.22.0
certifi==2022.9.14
charset-normalizer==2.1.1
configparser==5.3.0
frozenlist==1.3.1
fxcmpy==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
numpy==1.23.3
pandas==1.4.4
pydantic==1.10.2
Pygments==2.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.1
python-socketio==5.0.4
pytz==2022.2.1
requests==2.28.1
six==1.16.0
socketIO-client==0.7.2
socketIO-client-nexus==0.7.6
typing-extensions==4.3.0
urllib3==1.26.12
websocket-client==1.3.3
yarl==1.8.1
